LayoutTests/svg/animations/svgnumberoptionalnumber-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Test 'to' animation of SVGNumberOptionalNumber with optional number.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feGaussianBlur.stdDeviationX.animVal is 5
PASS feGaussianBlur.stdDeviationY.animVal is 5
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS feGaussianBlur.stdDeviationX.animVal is 22.5
PASS feGaussianBlur.stdDeviationY.animVal is 7.5
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS feGaussianBlur.stdDeviationX.animVal is 40
PASS feGaussianBlur.stdDeviationY.animVal is 10
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS feGaussianBlur.stdDeviationX.animVal is 5
PASS feGaussianBlur.stdDeviationY.animVal is 5
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS successfullyParsed is true

TEST COMPLETE
